+++ packages/core/src/actions/getConnectorClient.test.ts
@@ -86,20 +86,20 @@ test('behavior: account does not exist on connector', async () => {
 
 test('behavior: reconnecting', async () => {
   config.setState((state) => ({ ...state, status: 'reconnecting' }))
-  const { id, name, type, uuid } = connector
+  const { id, name, type, uid } = connector
   await expect(
     getConnectorClient(config, {
       connector: {
         id,
         name,
         type,
-        uuid,
+        uid,
       } as unknown as Connector,
     }),
   ).rejects.toThrowErrorMatchingInlineSnapshot(`
     [ConnectorUnavailableReconnectingError: Connector "Mock Connector" unavailable while reconnecting.
 
-    Details: During the reconnection step, the only connector methods guaranteed to be available are: \`id\`, \`name\`, \`type\`, \`uuid\`. All other methods are not guaranteed to be available until reconnection completes and connectors are fully restored. This error commonly occurs for connectors that asynchronously inject after reconnection has already started.
+    Details: During the reconnection step, the only connector methods guaranteed to be available are: \`id\`, \`name\`, \`type\`, \`uid\`. All other methods are not guaranteed to be available until reconnection completes and connectors are fully restored. This error commonly occurs for connectors that asynchronously inject after reconnection has already started.
     Version: @wagmi/core@x.y.z]
   `)
   config.setState((state) => ({ ...state, status: 'disconnected' }))