+++ packages/react/src/hooks/useConnect.ts
@@ -29,7 +29,7 @@ export type UseConnectParameters<
       | UseMutationParameters<
           ConnectData<config>,
           ConnectErrorType,
-          ConnectVariables<config>,
+          ConnectVariables<config, config['connectors'][number]>,
           context
         >
       | undefined
@@ -43,12 +43,12 @@ export type UseConnectReturnType<
   UseMutationReturnType<
     ConnectData<config>,
     ConnectErrorType,
-    ConnectVariables<config>,
+    ConnectVariables<config, config['connectors'][number]>,
     context
   > & {
     connect: ConnectMutate<config, context>
     connectAsync: ConnectMutateAsync<config, context>
-    connectors: Compute<UseConnectorsReturnType>
+    connectors: Compute<UseConnectorsReturnType> | config['connectors']
   }
 >
 
@@ -80,10 +80,11 @@ export function useConnect<
     )
   }, [config, result.reset])
 
+  type Return = UseConnectReturnType<config, context>
   return {
-    ...result,
-    connect: mutate,
-    connectAsync: mutateAsync,
+    ...(result as Return),
+    connect: mutate as Return['connect'],
+    connectAsync: mutateAsync as Return['connectAsync'],
     connectors: useConnectors({ config }),
   }
 }