+++ packages/vue/src/exports/index.ts
@@ -230,6 +230,8 @@ export {
   type ConnectorEventMap,
   type CreateConnectorFn,
   createConnector,
+  injected,
+  mock,
   // Errors
   type ChainNotConfiguredErrorType,
   ChainNotConfiguredError,